Reorganize code for columnformers and topomoe
Reorganize code for columnformers and topomoe into separate projects
with separate pyproject.toml, src and test directories, etc.
