+++ columnformers/src/extract_features.py
@@ -12,18 +12,21 @@ from typing import List, Optional
 
 import torch
 import yaml
-from timm.data import DEFAULT_CROP_PCT, create_dataset, create_loader
 from timm.utils import AverageMeter, random_seed
 from torch import Tensor
 from torch.utils.data import DataLoader
-from tqdm import tqdm
 from transformers.hf_argparser import HfArg, HfArgumentParser
 
-from topomoe import utils as ut
-from topomoe.inspection.features import FeatureExtractor, H5Writer, process_features
-from topomoe.models import create_model, list_models
-from topomoe.train import Args as TrainArgs
-from topomoe.train import get_num_classes, load_dataset_in_memory
+from columnformers import utils as ut
+from columnformers.data import create_dataset, create_loader, list_datasets
+from columnformers.inspection.features import (
+    FeatureExtractor,
+    H5Writer,
+    process_features,
+)
+from columnformers.models import create_model, list_models
+from columnformers.models.columnformer import AttnMode, MlpMode, NormMode
+from columnformers.train import _get_enum_values, parse_csv, to_device
 
 logging.basicConfig(
     format="[%(levelname)s %(asctime)s]: %(message)s", level=logging.INFO
@@ -42,27 +45,74 @@ class Args:
 
     # Model
     model: str = HfArg(
-        default="topomoe_tiny_2s_patch16_128",
+        default="vision_transformer_tiny_patch16_128",
         help=f"model ({', '.join(list_models())})",
     )
-    num_classes: Optional[int] = HfArg(default=None, help="number of classes")
+    attn_mode: Optional[str] = HfArg(
+        default=None,
+        help=f"attention mode ({', '.join(_get_enum_values(AttnMode))})",
+    )
+    mlp_mode: Optional[str] = HfArg(
+        default=None,
+        help=f"mlp mode ({', '.join(_get_enum_values(MlpMode))})",
+    )
+    norm_mode: Optional[str] = HfArg(
+        default=None,
+        help=f"norm mode ({', '.join(_get_enum_values(NormMode))})",
+    )
     num_heads: Optional[int] = HfArg(
         aliases=["--nh"], default=None, help="number of attention heads"
     )
-    mlp_ratio: Optional[List[float]] = HfArg(
+    mlp_ratio: Optional[str] = HfArg(
         aliases=["--mlpr"],
         default=None,
         help="mlp ratio. can be a single value or a list of values, "
-        "e.g. --mlpr 4.0, --mlpr 4.0 4.0 4.0 2.0 2.0 2.0",
+        "e.g. '4', '4,4,4,2,2,2'",
     )
-    num_experts: Optional[List[int]] = HfArg(
+    skip_attn: Optional[bool] = HfArg(
+        aliases=["--skip"], default=None, help="include attention skip connection"
+    )
+    attn_bias: Optional[bool] = HfArg(
+        aliases=["--attnb"], default=None, help="use learned attention bias"
+    )
+    attn_head_bias: Optional[bool] = HfArg(
+        aliases=["--attnhb"], default=None, help="use learned attention bias per head"
+    )
+    qk_head_dim: Optional[int] = HfArg(
+        aliases=["--qkd"], default=None, help="query and key head dimension"
+    )
+    no_vp: Optional[bool] = HfArg(
+        aliases=["--novp"], default=None, help="don't use value and projection"
+    )
+    num_experts: Optional[str] = HfArg(
         default=None,
-        help="number of experts. can be a single value or a list of values",
+        help="number of experts for MoE MLP. can be a single value or a list of values, "
+        "e.g. '2', '1,1,2,2,4,4'",
+    )
+    pool_stages: Optional[str] = HfArg(
+        default=None,
+        help="stages to apply pooling, branching, and double blocks in quadformer, "
+        "e.g. '2,4'",
     )
     mlp_conserve: Optional[bool] = HfArg(
         default=None,
         help="Divide params by num experts "
         "`expert_params = dim * mlp_ratio / num_experts`",
+    )
+    init_local_attn: Optional[bool] = HfArg(
+        aliases=["--initloc"], default=None, help="initialize with local attention bias"
+    )
+    depth_offset: int = HfArg(
+        aliases=["--offset"], default=2.0, help="distance offset between layers"
+    )
+    global_pool: str = HfArg(
+        aliases=["--pool"], default="avg", help="global pooling mode (avg, spatial)"
+    )
+    pos_embed: Optional[bool] = HfArg(
+        aliases=["--pos"], default=None, help="use position embedding"
+    )
+    time_embed: Optional[bool] = HfArg(
+        aliases=["--time_embed"], default=None, help="use time embedding"
     )
     drop_rate: float = HfArg(aliases=["--dr"], default=0.0, help="head dropout rate")
     proj_drop_rate: float = HfArg(
@@ -71,36 +121,23 @@ class Args:
     attn_drop_rate: float = HfArg(
         aliases=["--adr"], default=0.0, help="attention dropout rate"
     )
-    add_pos: bool = HfArg(
-        default=False, help="add stage position embedding to pooled input"
-    )
-    wiring_lambd: float = HfArg(
-        aliases=["--wlambd"], default=0.0, help="wiring length penalty"
-    )
-    wiring_sigma: float = HfArg(
-        aliases=["--wsigma"], default=2.0, help="wiring length radius stdev"
-    )
     # Dataset
     dataset: str = HfArg(
-        default="hfds/clane9/imagenet-100", help="timm-compatible dataset name"
+        aliases=["--ds"],
+        default="imagenet-100",
+        help=f"dataset ({', '.join(list_datasets())})",
     )
-    data_dir: Optional[str] = HfArg(default=None, help="dataset directory")
-    download: bool = HfArg(default=True, help="download dataset")
-    scale: List[float] = HfArg(
-        default_factory=lambda: [0.5, 1.0], help="image random crop scale"
-    )
-    ratio: List[float] = HfArg(
-        default_factory=lambda: [3 / 4, 4 / 3], help="image random crop ratio"
+    crop_min_scale: float = HfArg(
+        aliases=["--scale"], default=1.0, help="image random crop scale"
     )
     hflip: float = HfArg(default=0.0, help="hflip probability")
     color_jitter: Optional[float] = HfArg(
         aliases=["--jitter"], default=None, help="color jitter value"
     )
-    crop_pct: float = HfArg(default=DEFAULT_CROP_PCT, help="eval crop pct")
     workers: int = HfArg(aliases=["-j"], default=0, help="data loading workers")
     prefetch: bool = HfArg(default=True, help="use cuda prefetching")
     in_memory: bool = HfArg(
-        aliases=["--inmem"], default=False, help="keep dataset in memory"
+        aliases=["--inmem"], default=True, help="keep dataset in memory"
     )
     # Optimization
     epochs: int = HfArg(default=100, help="number of epochs")
@@ -108,17 +145,19 @@ class Args:
         aliases=["--bs"], default=256, help="batch size per replica"
     )
     # Paths
-    # Features will be written to out_dir / model
+    # Features will be written to:
+    #   out_dir / model
     out_dir: Path = HfArg(
-        default=Path("topomoe_features"), help="path to root output directory"
+        default=Path("features"), help="path to root output directory"
     )
-    checkpoint_path: Optional[Path] = HfArg(
+    # Logistics
+    checkpoint: Optional[str] = HfArg(
         aliases=["--ckpt"], default=None, help="checkpoint to load"
     )
-    model_args_path: Optional[Path] = HfArg(
-        aliases=["--args"], default=None, help="args to load for trained model"
-    )
     strict_load: bool = HfArg(aliases=["--strict"], default=True, help="strict loading")
+    restart: bool = HfArg(
+        default=False, help="Restart training rather than resume from checkpoint"
+    )
     cuda: bool = HfArg(default=True, help="use cuda")
     amp: bool = HfArg(default=False, help="use AMP")
     amp_dtype: str = HfArg(default="float16", help="AMP dtype (float16, bfloat16)")
@@ -156,74 +195,75 @@ def main(args: Args):
 
     logging.info(f"Loading split {args.split}")
 
+    input_size = int(args.model.split("_")[-1])
     dataset = create_dataset(
         args.dataset,
-        root=args.data_dir,
-        split=args.split,
-        is_training=True,
-        download=args.download,
-        batch_size=args.batch_size,
-    )
-    if args.in_memory:
-        logging.info("Loading dataset into memory")
-        load_dataset_in_memory(dataset)
-
-    input_size = int(args.model.split("_")[-1])
-    input_size = (3, input_size, input_size)
-
-    # TODO: check what the specific strings are for the splits in is_training
-    loader = create_loader(
-        dataset,
         input_size=input_size,
-        batch_size=args.batch_size,
-        is_training=True if args.split == "train" else False,
-        scale=args.scale,
-        ratio=args.ratio,
+        min_scale=args.crop_min_scale,
         hflip=args.hflip,
-        crop_pct=args.crop_pct,
         color_jitter=args.color_jitter,
-        interpolation="bicubic",
-        num_workers=args.workers,
-        persistent_workers=args.workers > 0,
-        distributed=clust.ddp,
-        device=clust.device,
-        use_prefetcher=args.prefetch,
+        keep_in_memory=args.in_memory,
     )
-    logging.info(f"\n\tnum samples: {len(dataset)}\n")
+    num_classes = dataset["train"].features["label"].num_classes
+
+    loaders = {}
+    for split, ds in dataset.items():
+        loaders[split] = create_loader(
+            ds,
+            shuffle=True,
+            batch_size=args.batch_size,
+            drop_last=True,
+            num_workers=args.workers,
+            distributed=clust.ddp,
+            pin_memory=not args.prefetch,
+            use_prefetcher=args.prefetch,
+            device=clust.device,
+        )
+    logging.info(f"\n\tnum samples: {len(ds)}\n")
 
     logging.info("Creating model %s", args.model)
 
-    num_classes = get_num_classes(dataset)
-
-    if args.checkpoint_path and args.model_args_path:
-        model_arg_parser = HfArgumentParser(TrainArgs)
-        (model_args,) = model_arg_parser.parse_yaml_file(yaml_file=args.model_args_path)
-    else:
-        model_args = args
-
     model = create_model(
-        model_args.model,
-        num_heads=model_args.num_heads,
-        mlp_ratio=model_args.mlp_ratio,
-        num_experts=model_args.num_experts,
-        mlp_conserve=model_args.mlp_conserve,
+        args.model,
+        attn_mode=args.attn_mode,
+        mlp_mode=args.mlp_mode,
+        norm_mode=args.norm_mode,
+        num_heads=args.num_heads,
+        mlp_ratio=parse_csv(args.mlp_ratio, float),
+        skip_attn=args.skip_attn,
+        attn_bias=args.attn_bias,
+        attn_head_bias=args.attn_head_bias,
+        qk_head_dim=args.qk_head_dim,
+        no_vp=args.no_vp,
+        num_experts=parse_csv(args.num_experts),
+        pool_stages=parse_csv(args.pool_stages, squeeze=False),
+        mlp_conserve=args.mlp_conserve,
+        init_local_attn=args.init_local_attn,
+        depth_offset=args.depth_offset,
         num_classes=num_classes,
-        drop_rate=model_args.drop_rate,
-        proj_drop_rate=model_args.proj_drop_rate,
-        attn_drop_rate=model_args.attn_drop_rate,
-        add_pos=model_args.add_pos,
-        wiring_lambd=model_args.wiring_lambd,
-        wiring_sigma=model_args.wiring_sigma,
+        global_pool=args.global_pool,
+        pos_embed=args.pos_embed,
+        time_embed=args.time_embed,
+        drop_rate=args.drop_rate,
+        proj_drop_rate=args.proj_drop_rate,
+        attn_drop_rate=args.attn_drop_rate,
     )
     model: torch.nn.Module = model.to(device)
     logging.info(
         f"param count: {sum([m.numel() for m in model.parameters()])/1e6:.0f}M"
     )
 
-    if args.checkpoint_path:
-        logging.info("Loading checkpoint: %s", args.checkpoint_path)
-        ckpt = torch.load(args.checkpoint_path, map_location=device)
-        model.load_state_dict(ckpt["model"])
+    if args.checkpoint:
+        optimizer = ut.create_optimizer(model)
+        logging.info("Loading checkpoint: %s", args.checkpoint)
+        ut.load_checkpoint(
+            args.checkpoint,
+            model,
+            optimizer,
+            device=clust.device,
+            strict=args.strict_load,
+            load_opt_state=not args.restart,
+        )
 
     # Construct feature extractor
     # NOTE: there is also torchvision create_feature_extractor() we may want to
@@ -237,7 +277,7 @@ def main(args: Args):
     )
 
     logging.info(f"Extracting to {out_path}")
-    extract_features(args, extractor, loader, out_path, device)
+    extract_features(args, extractor, loaders[args.split], out_path, device)
 
     runtime = time.monotonic() - start_time
     logging.info(f"Done! runtime: {runtime:.2f}s")
@@ -251,7 +291,6 @@ def extract_features(
     out_path: Path,
     device: torch.device,
 ) -> Tensor:
-
     extractor.model.eval()
 
     num_samples = len(loader.dataset)
@@ -262,16 +301,13 @@ def extract_features(
     last_idx = len(loader) - 1
     with writer as writer:
         for epoch in range(args.epochs):
-            for batch_idx, (input, target) in tqdm(
-                enumerate(loader), total=len(loader)
-            ):
+            for batch_idx, batch in enumerate(loader):
                 last_batch = batch_idx == last_idx
-
-                if not args.prefetch:
-                    input, target = input.to(device), target.to(device)
+                batch = to_device(batch, device)
+                images = batch["image"]
 
                 # Each feature shape: (batch_size, ...)
-                _, batch_features = extractor(input)
+                _, batch_features = extractor(images)
 
                 for name, values in batch_features.items():
                     # Reshape features to (N, T, D) and optionally perform adaptive
