+++ columnformers/src/inspection/figures.py
