+++ columnformers/src/inspection/registry.py
