+++ columnformers/src/models/geometry.py
