+++ columnformers/src/tests/test_inspection/test_figures.py
