+++ columnformers/src/tests/test_models/test_vision_columnformer.py
