+++ topomoe/src/inspection/most_activating_images.py
