+++ topomoe/src/utils/misc.py
